Modify code to use the changed adapter behaviour
